[BeverageDetails] Show Notes if user is logged in. Minor fix of HopRate schema and Stats data normalizer

+++ gatsby-node.js
@@ -251,7 +251,7 @@ exports.createSchemaCustomization = ({ actions, schema }) => {
         producer: HopRateValues
       }
       type HopRateValues {
-        value: Int!
+        value: Float!
 			  unit: HopRateUnitEnum!
       }
       enum HopRateUnitEnum {
